+++ src/instructlab/model/backends/backends.py
@@ -315,12 +315,14 @@ def get_uvicorn_config(app: fastapi.FastAPI, host: str, port: int) -> Config:
     )
 
 
-def select_backend(cfg: serve_config) -> BackendServer:
+def select_backend(
+    cfg: serve_config, model_path: pathlib.Path | None = None
+) -> BackendServer:
     # Local
     from .llama_cpp import Server as llama_cpp_server
     from .vllm import Server as vllm_server
 
-    model_path = pathlib.Path(cfg.model_path)
+    model_path = pathlib.Path(model_path or cfg.model_path)
     backend_name = cfg.backend
     try:
         backend = get(model_path, backend_name)